# Improve performance for SawyerChunkReader once again