--- a/src/openrct2/rct12/SawyerChunk.cpp
+++ b/src/openrct2/rct12/SawyerChunk.cpp
@@ -9,10 +9,11 @@
 
 #include "SawyerChunk.h"
 
-#include "../core/Memory.hpp"
 #include "SawyerChunkReader.h"
 
-SawyerChunk::SawyerChunk(SAWYER_ENCODING encoding, std::vector<uint8_t>&& data)
+using namespace OpenRCT2;
+
+SawyerChunk::SawyerChunk(SAWYER_ENCODING encoding, MemoryStream&& data)
     : _data(std::move(data))
     , _encoding(encoding)
 {