--- a/src/openrct2/rct12/SawyerChunkReader.cpp
+++ b/src/openrct2/rct12/SawyerChunkReader.cpp
@@ -15,16 +15,6 @@
 
 using namespace OpenRCT2;
 
-// malloc is very slow for large allocations in MSVC debug builds as it allocates
-// memory on a special debug heap and then initialises all the memory to 0xCC.
-#if defined(_WIN32) && defined(DEBUG)
-#    define __USE_HEAP_ALLOC__
-#    ifndef WIN32_LEAN_AND_MEAN
-#        define WIN32_LEAN_AND_MEAN
-#    endif
-#    include <windows.h>
-#endif
-
 // Allow chunks to be uncompressed to a maximum of 16 MiB
 constexpr size_t MAX_UNCOMPRESSED_CHUNK_SIZE = 16 * 1024 * 1024;