This drastically improves the load times for debug builds and squeezed a few ms out of release builds.
## Debug
develop:
```
Building object index (7929 items)
Finished building object index in 88.55 seconds.
```
PR:
```
Building object index (7929 items)
Finished building object index in 7.49 seconds.
```

## Release
develop:
```
Building object index (7929 items)
Finished building object index in 0.39 seconds.
```
PR:
```
Building object index (7929 items)
Finished building object index in 0.36 seconds.
```